+++ main.h
@@ -49,7 +49,7 @@ void freeGraph(Graph *g);
 void etapaInitiala(Node *head, Node **castigatori, Node **pierzatori, Graph *g);
 void etapaFinala(Node **castigatori, Node **pierzatori, Graph *g);
 void introduceInGraf(Graph *g, Node *echipa_castigatoare, Node *echipa_pierzatoare);
-double calculPrestigiu(Node *echipa);
+float calculPrestigiu(Node *echipa);
 void afisarePrestigiuEchipe(FILE* fisier_out_scor, Node *castigatori, Node *pierzatori);
 
 void Cerinta_1(FILE* fisier_out_graf, Graph *g);
